+++ ops.sh
@@ -108,7 +108,7 @@ ops-composer() {
 
     local project="$(ops project name)"
 
-    ops docker run \
+    _ops-docker run \
         --rm -itP \
         -v "$(pwd):/var/www/html/$project" \
         -v "$HOME/.composer:/var/www/.composer" \
@@ -137,7 +137,7 @@ ops-exec() {
 
     [[ -z $id ]] && exit
 
-    ops docker exec -i $id "$@"
+    _ops-docker exec -i $id "$@"
 }
 
 ops-help--after() {
@@ -222,15 +222,23 @@ mariadb-create() {
     fi
 }
 
+mariadb-list() {
+    ops-exec mariadb mysql --column-names=FALSE -e "show databases;" | \
+        grep -v "^information_schema$" | \
+        grep -v "^performance_schema$" | \
+        grep -v "^mysql$"
+}
+
 mariadb-export() {
     local db="$1"
 
-    ops-exec mariadb mysqldump --single-transaction "$db"
+    ops-exec mariadb mysqldump --single-transaction --add-drop-table "$db"
 }
 
 mariadb-import() {
     local db="$1"
     local sqlfile=${2--}
+
 
     (
         # don't let these commands grab stdin
@@ -322,6 +330,11 @@ psql-run() {
     ops-exec postgres psql -U postgres "${@}"
 }
 
+psql-list() {
+    ops-exec postgres psql -U postgres -t -c "SELECT datname FROM pg_database WHERE datname NOT IN ('template0', 'template1', 'postgres')" | \
+    sed -e "s/^ *//" -e "/^$/d"
+}
+
 psql-export() {
     local db="$1"
 
@@ -352,20 +365,21 @@ ops-psql() {
     cmd-run psql "$@"
 }
 
-_ops-lt() {
+ops-lt() {
+    cmd-doc "Create a localtunnel to your project"
+
     local project="$(ops project name)"
 
-    echo "$project.$OPS_DOMAIN"
+    echo "Creating a localtunnel to $project.$OPS_DOMAIN"
+    echo
 
     ops docker run \
         --rm --init -itP \
         --label=ops.project="$project" \
-        --network=host \
-        efrecon/localtunnel \
-            --local-host="$project.$OPS_DOMAIN" \
-            --port=80
-
-        #ops-utils:$OPS_VERSION bash \
+        --network="ops_services" \
+        --dns="$OPS_SERVICES_DNS_IP" \
+        imarcagency/ops-localtunnel:latest \
+           --host="$OPS_LOCALTUNNEL_HOST" --local-host="$project.$OPS_DOMAIN" --port=80
 }
 
 _ops-gulp() {
@@ -390,20 +404,16 @@ ops-restart() {
     cmd-doc "Restart all running containers"
     ops-stop
 
-
     ops-start
 }
 
 ops-shell() {
     cmd-doc "Enter shell or execute command"
 
-    local id=$(system-docker-compose ps -q $OPS_SHELL_BACKEND)
+    local id=$(system-docker-compose ps -q $OPS_SHELL_BACKEND 2> /dev/null)
+    local project_id=$(project-docker-compose ps -q $OPS_SHELL_BACKEND 2> /dev/null)
     local project=$(project-name)
     local command="$OPS_SHELL_COMMAND"
-
-    if [[ -z $id ]]; then
-        exit
-    fi
 
     if [[ ! -z "$1" ]]; then
         command="$@"
@@ -414,7 +424,15 @@ ops-shell() {
         t='t'
     fi
 
-    ops docker exec -w "/var/www/html/$project" -u "$OPS_SHELL_USER" -i$t $id $command
+    if [[ ! -z $id ]]; then
+        _ops-docker exec -w "/var/www/html/$project" -u "$OPS_SHELL_USER" -i$t $id $command
+    elif [[ ! -z $project_id ]]; then
+        _ops-docker exec -u "$OPS_SHELL_USER" -i$t $project_id $command
+    else
+        echo "ERROR: No such service: $OPS_SHELL_BACKEND"
+        exit 1
+
+    fi
 }
 
 ops-link() {
@@ -462,7 +480,7 @@ ops-stats() {
         exit
     fi
 
-    ops docker stats $ids
+    _ops-docker stats $ids
 }
 
 ops-start() {
@@ -501,7 +519,7 @@ ops-stop() {
 
     system-stop
 
-    local info=$(ops docker ps -a --format '{{.ID}} {{.Label "ops.project"}}' --filter="label=ops.project")
+    local info=$(_ops-docker ps -a --format '{{.ID}} {{.Label "ops.project"}}' --filter="label=ops.project")
 
     IFS=$'\n'
     for container in $info; do
@@ -551,7 +569,7 @@ ops-sync() {
     fi
 
     cd "$OPS_SITES_DIR/$OPS_PROJECT_NAME"
-        #source ".env"
+    #source ".env"
 
     # best debugging helper
     # ( set -o posix ; set ) | grep -E '^OPS_'
@@ -575,10 +593,12 @@ ops-sync() {
         [[ ! -z "$OPS_PROJECT_REMOTE_DB_TYPE" ]] && \
         [[ ! -z "$OPS_PROJECT_REMOTE_DB_NAME" ]]
     then
-        if [[ "$OPS_PROJECT_REMOTE_OPS" ]]; then
+        if [[ "$OPS_PROJECT_REMOTE_OPS" != 0 ]]; then
+            echo "Syncing remote mariadb '$OPS_PROJECT_REMOTE_DB_NAME' to local '$OPS_PROJECT_DB_NAME'"
+
             ssh -C "$ssh_host" \
                 "ops $OPS_PROJECT_REMOTE_DB_TYPE export $OPS_PROJECT_REMOTE_DB_NAME" | \
-                ops $OPS_PROJECT_DB_TYPE import $OPS_PROJECT_DB_NAME
+                $OPS_PROJECT_DB_TYPE-import "$OPS_PROJECT_DB_NAME"
 
         elif [[ "$OPS_PROJECT_REMOTE_DB_TYPE" = "mariadb" ]]; then
             echo "Syncing remote mariadb '$OPS_PROJECT_REMOTE_DB_NAME' to local '$OPS_PROJECT_DB_NAME'"
@@ -729,7 +749,7 @@ project-name() {
 }
 
 project-start() {
-    project-docker-compose up -d "$@"
+    project-docker-compose up -d --force-recreate "$@"
 }
 
 project-stop() {
@@ -754,7 +774,7 @@ project-exec() {
         exit
     fi
 
-    ops docker exec -i $id "$@"
+    _ops-docker exec -i $id "$@"
 }
 
 project-help() {
@@ -770,7 +790,7 @@ project-shell-exec() {
         exit
     fi
 
-    ops docker exec -it $id "$@"
+    _ops-docker exec -it $id "$@"
 }
 
 project-stats() {
@@ -780,7 +800,7 @@ project-stats() {
         exit
     fi
 
-    ops docker stats $ids
+    _ops-docker stats $ids
 }
 
 # System Sub-Commands
@@ -819,7 +839,7 @@ system-shell-exec() {
         exit 1
     fi
 
-    ops docker exec -it $id "$@"
+    _ops-docker exec -it $id "$@"
 }
 
 system-check() {
@@ -906,9 +926,9 @@ system-install() {
         mkdir -p $OPS_HOME/certs
 
         rsync -a \
-          --exclude=bin \
-          --exclude=certs \
-          --exclude=config \
+          --exclude=/bin \
+          --exclude=/certs \
+          --exclude=/config \
           --exclude=acme.json \
           $OPS_SCRIPT_DIR/home/ \
           $OPS_HOME
@@ -919,7 +939,6 @@ system-install() {
     source $OPS_HOME/config
 
     system-install-mkcert
-    system-refresh-config
 
     if [[ ! -f "$OPS_HOME/certs/self-signed-cert.key" ]]; then
         system-refresh-certs
@@ -997,19 +1016,6 @@ system-refresh-certs() {
 
 }
 
-system-refresh-config() {
-    #
-    # Build config
-    #
-
-    sed "s/OPS_DOMAIN/$OPS_DOMAIN/" $OPS_HOME/dnsmasq/dnsmasq.conf.tmpl > $OPS_HOME/dnsmasq/dnsmasq.conf
-
-    sed \
-        -e "s/OPS_MINIO_ACCESS_KEY/$OPS_MINIO_ACCESS_KEY/" \
-        -e "s/OPS_MINIO_SECRET_KEY/$OPS_MINIO_SECRET_KEY/" \
-        $OPS_HOME/minio/config.json.tmpl > $OPS_HOME/minio/config.json
-}
-
 system-refresh-services() {
     #
     # Regenerate/Restart services. (They might depend on new configs/certs)
@@ -1034,10 +1040,29 @@ system-reset() {
     _ops-docker network rm ops_backend &> /dev/null
     _ops-docker network rm ops_gateway &> /dev/null
     _ops-docker network rm ops_services &> /dev/null
+
+    # TODO inspect networks and point
+    # docker network inspect -f '{{range .IPAM.Config}}{{ $.Name }} {{.Subnet}}{{end}}' $(docker network ls -q) | sed '/^$/d'
+
+    # create networks
+    _ops-docker network create --subnet="$OPS_SERVICES_SUBNET" ops_services &> /dev/null
+    _ops-docker network create ops_backend &> /dev/null
+    _ops-docker network create ops_gateway &> /dev/null
 }
 
 system-start() {
     system-reset
+
+    # refresh config
+    sed \
+        -e "s/OPS_DOMAIN/$OPS_DOMAIN/" \
+        -e "s/OPS_SERVICES_TRAEFIK_IP/$OPS_SERVICES_TRAEFIK_IP/" \
+        $OPS_HOME/dnsmasq/dnsmasq.conf.tmpl > $OPS_HOME/dnsmasq/dnsmasq.conf
+
+    sed \
+        -e "s/OPS_MINIO_ACCESS_KEY/$OPS_MINIO_ACCESS_KEY/" \
+        -e "s/OPS_MINIO_SECRET_KEY/$OPS_MINIO_SECRET_KEY/" \
+        $OPS_HOME/minio/config.json.tmpl > $OPS_HOME/minio/config.json
 
     # start all services
     system-docker-compose up -d --remove-orphans
@@ -1062,6 +1087,9 @@ main() {
 
         validate-config
     fi
+
+    declare -x OPS_SERVICES_DNS_IP="$(echo -n $OPS_SERVICES_SUBNET | cut -f1,2 -d'.' | sed 's/$/.10.10/')"
+    declare -x OPS_SERVICES_TRAEFIK_IP="$(echo -n $OPS_SERVICES_SUBNET | cut -f1,2 -d'.' | sed 's/$/.10.11/')"
 
     cmd-run ops "$@"
     exit
@@ -1088,9 +1116,9 @@ fi
 
 declare -x OPS_ENV="dev"
 declare -x OPS_DEBUG="${OPS_DEBUG}"
-declare -x OPS_BACKENDS=${OPS_BACKENDS-"apache-php73"}
+declare -x OPS_BACKENDS=${OPS_BACKENDS-"apache-php74"}
 declare -x OPS_SERVICES=${OPS_SERVICES-"dnsmasq portainer dashboard mariadb postgres redis adminer"}
-declare -x OPS_DOCKER_COMPOSER_IMAGE=${OPS_DOCKER_COMPOSER_IMAGE-"imarcagency/ops-apache-php73:$OPS_VERSION"}
+declare -x OPS_DOCKER_COMPOSER_IMAGE=${OPS_DOCKER_COMPOSER_IMAGE-"imarcagency/ops-apache-php74:$OPS_VERSION"}
 declare -x OPS_DOCKER_NODE_IMAGE=${OPS_DOCKER_NODE_IMAGE-"imarcagency/ops-node:$OPS_VERSION"}
 declare -x OPS_DOCKER_UTILS_IMAGE=${OPS_DOCKER_UTILS_IMAGE-"imarcagency/ops-utils:$OPS_VERSION"}
 declare -x OPS_DOCKER_GID=${OPS_DOCKER_GID-""}
@@ -1101,16 +1129,18 @@ declare -x OPS_DOMAIN=${OPS_DOMAIN-"imarc.io"}
 declare -x OPS_MINIO_ACCESS_KEY=${OPS_MINIO_ACCESS_KEY-"minio-access"}
 declare -x OPS_MINIO_SECRET_KEY=${OPS_MINIO_SECRET_KEY-"minio-secret"}
 declare -x OPS_SITES_DIR=${OPS_SITES_DIR-"$HOME/Sites"}
+declare -x OPS_SERVICES_SUBNET=${OPS_SERVICES_SUBNET-"172.23.0.0/16"}
 declare -x OPS_ACME_EMAIL=${OPS_ACME_EMAIL-""}
 declare -x OPS_ACME_DNS_PROVIDER=${OPS_ACME_DNS_PROVIDER-""}
 declare -x OPS_ACME_PRODUCTION=${OPS_ACME_PRODUCTION-"0"}
 declare -x OPS_ADMIN_AUTH=${OPS_ADMIN_AUTH-""}
 declare -x OPS_ADMIN_AUTH_LABEL_PREFIX=""
+declare -x OPS_LOCALTUNNEL_HOST=${OPS_LOCALTUNNEL_HOST-"https://localtunnel.me"}
 
-declare -x OPS_DEFAULT_BACKEND=${OPS_DEFAULT_BACKEND-"apache-php73"}
+declare -x OPS_DEFAULT_BACKEND=${OPS_DEFAULT_BACKEND-"apache-php74"}
 declare -x OPS_DEFAULT_DOCROOT=${OPS_DEFAULT_DOCROOT-"public"}
 declare -x OPS_DASHBOARD_URL="https://ops.${OPS_DOMAIN}"
-declare -x OPS_MKCERT_VERSION="1.3.0"
+declare -x OPS_MKCERT_VERSION="1.4.1"
 
 if [[ ! $OPS_ADMIN_AUTH ]]; then
     OPS_ADMIN_AUTH_LABEL_PREFIX="disabled-"
@@ -1144,6 +1174,7 @@ declare -x OPS_PROJECT_SYNC_DIRS="${OPS_PROJECT_SYNC_DIRS}"
 declare -x OPS_PROJECT_SYNC_NODB="${OPS_PROJECT_SYNC_NODB-0}"
 declare -x OPS_PROJECT_SYNC_EXCLUDES="${OPS_PROJECT_SYNC_EXCLUDES}"
 declare -x OPS_PROJECT_SYNC_MAXSIZE="${OPS_PROJECT_SYNC_MAXSIZE}"
+declare -x OPS_PROJECT_REMOTE_OPS="${OPS_PROJECT_REMOTE_OPS-0}"
 declare -x OPS_PROJECT_REMOTE_USER="${OPS_PROJECT_REMOTE_USER}"
 declare -x OPS_PROJECT_REMOTE_HOST="${OPS_PROJECT_REMOTE_HOST}"
 declare -x OPS_PROJECT_REMOTE_PATH="${OPS_PROJECT_REMOTE_PATH}"